use np.float32, test=develop

+++ python/paddle/fluid/tests/unittests/test_queue.py
@@ -67,7 +67,7 @@ class TestQueue(unittest.TestCase):
         exe.run(startup_program)
         ret = exe.run(main_program, fetch_list=[data_out.name])
         self.assertTrue(
-            np.allclose(np.asarray(ret), np.full((3, 2), value, np.int32)))
+            np.allclose(np.asarray(ret), np.full((3, 2), value, np.float32)))
 
 
 if __name__ == '__main__':
